genericize control flow analyses and context info
control flow analyses include a possible callback to do per-instruction
analysis when discovered through control flow exploration. it's not
anticipated that instructions will be discovered without control flow
analysis being run, but this may change in the future

to support a per-instruction callback including information updates,
ContextTable grew a `put` method to update information retained.
Implementors currently only support updating an entire computed context
as a unit, but this is anticipated to grow in the future to partial
context updates ("we know r5 == 0x8" as a granular piece), possibly
serving as the backbone for update-triggered analyses?