add supporting logic for operand scrolling and location highlights
arguably this demonstrates how much of -core should be either feature
flagged behind `display` or in a second inner crate focused entirely on
instruction layout and styling. this commit adds a struct to -core for
operand selection, and location selection in operands, and a skeleton of
implementation for x86_64. it also adds a trait which is impl'd for
instruction highlighting rules, and most default uses just pass a struct
that indicates no highlights anywhere. for display use, this might be
backed by a Vec of addresses to indicate highlighting, or more precise
selectors for operand highlighting as well.

this all spiders through into needing changes to -core x86_64
instruction display. it's not great that this much stuff is involved for
what might be dead simple rendering in dependencies. it might be
plausible to move even more stuff to trait-bounded type params so that
specializations of these functions for minimal use can just drop most
cruft via DCE