add a notion of default abi, and function vs function implementation
function implementations are the manifestation of a function as seen in
a real program, with respect to some abi. this abi may not be the
default (consider: windows 32bit abi is not the same as fastcall, rust
abi often does match but may not match the target abi for internal
functions)

outstanding things to do: point to the actual abi used for a function,
enumerating the abis used by a program, being able to select a new abi
in the case an inferred abi doesn't match the real use for a function.
declaring new parameters and return values for functions and their
types