fix issue where reads after a read/write record a false read
in instructions like "xor rax, rax" include a read/write use of the
first operand, followed by a read of the second operand. when computing
a value numbering for these registers, the first read causes a read of
the extant rax, followed by a write to declare a new rax, but the last
read then is interpreted as a read of the new rax, rather than a read of
the original rax.

this also updates the address/location/direction mapping to the value
that's read, causing later lookups of versions of registers read to
yield the newly-written value as well.

because instructions are assumed to be atomic with respect to CPU (and
other architecture details') state, it seems a safe assumption that any
read in an instruction references state before the current instruciton,
and any write declares state as modified by the current instruction.
this commit introduces a set to track if a value has been written yet,
to smooth out value relations in the face of an instruciton's locations
consisting of a sequence of mixed reads and writes to possibly the same
location.