add a notion of location disambiguation and extend x86_64 to use it
this is intended to begin support of precise memory analyses, and
revertable assumptions about program and machine layout. these
assumptions are things like "stack, heap, static, and other memory
accesses do not alias one another, and stack accesses are either
consistently through the stack pointer (or an sp-derived value), or
through some other value pointing into the stack, but not
interchangably".

the last point about interchangable access is what makes assuming a
no-clobber of locals when passed a stack-referncing pointer to a struct
plausible. otherwise we would have to involve points-to analysis and
reconsider at every call site. we may still have to to autonomously
determinte when this assumption is violated.