avoid panic when rendering instructions at the end of a memory region
because memory ranges are inclusive, the range [start] + [instr.len()]
is a range one element longer than the number of bytes in the
instruction. when bounds-checking this range against the underlying
memory, this will be too large, and yield None, which we then unwrap to
display.

solution: just use a range starting at `start`, since we iterate only
`instr.len()` elements to show instruction bytes