control flow analysis optimizations
this contains two algorithmic corrections and a smaller poorly-tested
optimization.

* first, `ControlFlowGraph::get_function` attempted to avoid following
tail calls by marking functions as basic blocks that have already been
seen. this has the unfortunate result of iterating all functions and
collecting into the `graphmap`'s `Bfs` iterator. as a consequence, as
a program's basic blocks are more fully understood, each call to
`get_functions` takes on overhead linear in the number of basic blocks.

since in typical programs the number of basic blocks is a function of
the number of functions, this seems to make the sum time of
`get_function` quadradic in the number of functons, rather than linear.

* second, `ControlFlowGraph.blocks` was a `Vec<BasicBlock<A>>` sorted on
the start of basic blocks. this was an important property as
conceptually a program had an address space from the lowest to highest
address of its architecture, with splits introduced as basic block
boundaries are found. indices of basic blocks to split were found with
an nlogn binary search, but i had neglected to consider that inserting
the split-off basic block would entail moving the *rest* of the Vec up
in memory. this makes for an expected number of copies linear with the
number of basic blocks, on every split.

because the number of splits matches the number of new basic blocks per
function, walking a new function has overhead like

```
|all_basic_blocks| * |basic_blocks_per_fn| * |number_of_functions|
```

again, since the number of basic blocks in a program is typically a
constantish multiple of the number of functions in a program, this makes
"walk all control flow in the program" roughly quadradic in the number
of functions.

* third, replace `Vec` in `ControlFlowGraph::with_effect` with `SmallVec`

this isn't strictly backed by any measurements, but `with_effect`
currently returns Vec of zero, one, or two items. when handling
multi-target branches, this may be an arbitrary number of items, but is
typically an uncommon case. this avoids allocating and dropping a Vec
for every instruction with control flow implications.