as seen on twitter dot com, "implementation of `DFGLocationQuery` is not general enough"
this introduces a notion of "indirect" queries to a DFG. "indirect"
describes the kinds of values backing memory and other arbitrarily
indexable and aliasing locations in an architecture. "Indirect" is an
associated type on a DFG because the specific details of a DFG's support
for indirect queries are a function of the values the DFG describes.
some DFGs, such as those for control flow effects, are entirely
acceptable in treating any indirection as an opaque unknown - for
control flow analyses it is likely acceptable for memory accesses to
yield unknown values, and for memory writes to be ignored. ignoring
memory writes, then, allows control flow effects to avoid unnecessary
work in tracking data that will never impact outputs.

however, by making `Indirect` itself generic on the DFG, if an
`Indirect` needs to hold a lifetime (for example, itself holding a
borrow of the `DFG` it is holding the curried arguments for, or a borrow
of the cursor it's adding an argument to), that would require writing
the associated type
```
trait IndirectDFG<V, A, W> {
    type Indirect<'dfg>: Indirect<'dfg, V>;
}
```

but rust does not have GAT, so such an associated type is not
permissible.

the other option to get a lifetime is to implement a trait providing
indirection queries/currying on a reference directly, which is where
`src/analyses/memory_layout.rs`' `impl IndirectDFG for &'cursor mut
MemoryLayout` comes in. the idea is that `MemoryLayout` is, itself, a
DFG over values that can be tracked in memory. indirection queries get a
`MemoryRegion2` (name pending) which associates an SSA'd location
supporting indirection with supporting memory layout information (in
theory, not reality). a `MemoryRegion2` then holds a borrow from the
cursor it was created from, which here also holds a borrow of `ssa`.
since any `DFG` implementor should support even a stub `indirect` that
treats indirection opaquely, `trait DFG` requires
```
for<'dfg> &'dfg mut Self: IndirectDFG
```
so implementations like
```
impl<'ssa, cursor> IndirectDFG for &'cursor mut MemoryLayout<'ssa, amd64>
```
can introduce a new arbitrary 'cursor lifetime to use for the associated
`type Indirect`.

unfortunately, somewhere along the way this makes an impl not generic
enough, and things fall over.