memory accesses through Alias expressions match referents
there was no mechanism to go from a Value { Data::Alias(v) } to the
underlying "v", so memory accesses would be present for both "v" and
registers that alias it. elsewhere with more canonicalization this means
accesses for "v" were missed, with the dfg tracking some "v" alias
instead. this change more aggressively canonicalizes memory accesses,
with a goal that the DFG backing some analyses always uses the
underlying aliases.